\name{closest.top}
\alias{closest.top}
\title{Find the 'top' closest gene}
\description{
Find the 'top' closest genes to the target gene based on a distance object 
}
\usage{
closest.top(x, dist.mat, top)
}
\arguments{
  \item{x}{ gene name }
  \item{dist.mat}{ calculated distance object }
  \item{top}{ number of closest genes desired }
}
\details{
}
\value{
 vector of top gene names
}
\references{ }
\author{ Beiying Ding }
\note{  }

\seealso{ \code{\link{cor.dist}}, \code{link{spearman.dist}}, \code{\link{tau.dist}},\code{\link{euc}},
\code{\link{man}},\code{\link{KLdist.matrix}},\code{\link{KLD.matrix}},\code{\link{mutualInfo}} } 
\examples{
 ##none yet
}
\keyword{}
